+++ pkgs/ffigen/lib/src/code_generator/func.dart
@@ -44,12 +44,12 @@ import 'writer.dart';
 /// ```
 class Func extends LookUpBinding with HasLocalScope {
   final FunctionType functionType;
-  final bool exposeSymbolAddress;
+  bool exposeSymbolAddress;
   final bool exposeFunctionTypedefs;
-  final bool isLeaf;
+  bool isLeaf;
   final bool objCReturnsRetained;
   final bool useNameForLookup;
-  final bool recordUse;
+  bool recordUse;
   final ApiAvailability? apiAvailability;
 
   @override