+++ pkgs/ffigen/lib/src/code_generator/global.dart
@@ -27,7 +27,7 @@ import 'writer.dart';
 /// ```
 class Global extends LookUpBinding with HasLocalScope {
   final Type type;
-  final bool exposeSymbolAddress;
+  bool exposeSymbolAddress;
   final bool constant;
   final ConstantValue? constantValue;