+++ pkgs/ffigen/lib/src/config_provider/yaml_config.dart
@@ -1263,44 +1263,18 @@ final class YamlConfig {
               ),
       ),
       functions: Functions(
-        includeSymbolAddress: functionDecl.shouldIncludeSymbolAddress,
         varArgs: varArgFunctions,
         includeTypedef: shouldExposeFunctionTypedef,
-        isLeaf: isLeafFunction,
       ),
-      structs: Structs(
-        dependencies: _structDependencies,
-        packingOverride: (decl) =>
-            _structPackingOverride.getOverridenPackValue(decl.originalName),
-      ),
-      enums: Enums(
-        silenceWarning: silenceEnumWarning,
-        style: (e, suggestedStyle) {
-          if (suggestedStyle != null) return suggestedStyle;
-          return switch (enumShouldBeInt(e)) {
-            true => EnumStyle.intConstants,
-            false => EnumStyle.dartEnum,
-          };
-        },
-      ),
-      unions: Unions(dependencies: _unionDependencies),
-      globals: Globals(
-        includeSymbolAddress: globals.shouldIncludeSymbolAddress,
-      ),
-      typedefs: Typedefs(
-        useSupportedTypedefs: useSupportedTypedefs,
-        includeUnused: includeUnusedTypedefs,
-      ),
+      typedefs: Typedefs(includeUnused: includeUnusedTypedefs),
       importType: importType,
       objectiveC: language == Language.objc
           ? ObjectiveC(
               interfaces: Interfaces(
                 includeTransitive: includeTransitiveObjCInterfaces,
-                module: interfaceModule,
               ),
               protocols: Protocols(
                 includeTransitive: includeTransitiveObjCProtocols,
-                module: protocolModule,
               ),
               categories: Categories(
                 includeTransitive: includeTransitiveObjCCategories,
@@ -1331,6 +1305,12 @@ final class YamlConfigAstVisitor extends public_ast.Visitor {
     if (config.functionDecl.rename(_decl(node)) case final rename?) {
       node.name = rename;
     }
+    if (config.isLeafFunction(_decl(node))) {
+      node.isLeaf = true;
+    }
+    if (config.functionDecl.shouldIncludeSymbolAddress(_decl(node))) {
+      node.exposeSymbolAddress = true;
+    }
   }
 
   @override
@@ -1347,6 +1327,10 @@ final class YamlConfigAstVisitor extends public_ast.Visitor {
     if (config.structDecl.rename(_decl(node)) case final rename?) {
       node.name = rename;
     }
+    if (config.structPackingOverride(_decl(node)) case final override?) {
+      node.pack = override.value;
+    }
+    node.dependencies = config.structDependencies;
   }
 
   @override
@@ -1355,6 +1339,7 @@ final class YamlConfigAstVisitor extends public_ast.Visitor {
     if (config.unionDecl.rename(_decl(node)) case final rename?) {
       node.name = rename;
     }
+    node.dependencies = config.unionDependencies;
   }
 
   @override
@@ -1363,6 +1348,10 @@ final class YamlConfigAstVisitor extends public_ast.Visitor {
     if (config.enumClassDecl.rename(_decl(node)) case final rename?) {
       node.name = rename;
     }
+    if (config.enumShouldBeInt(_decl(node))) {
+      node.style = EnumStyle.intConstants;
+    }
+    node.silenceWarning = config.silenceEnumWarning;
   }
 
   @override
@@ -1370,6 +1359,9 @@ final class YamlConfigAstVisitor extends public_ast.Visitor {
     node.isIncluded = config.globals.shouldInclude(_decl(node));
     if (config.globals.rename(_decl(node)) case final rename?) {
       node.name = rename;
+    }
+    if (config.globals.shouldIncludeSymbolAddress(_decl(node))) {
+      node.exposeSymbolAddress = true;
     }
   }
 
@@ -1395,6 +1387,9 @@ final class YamlConfigAstVisitor extends public_ast.Visitor {
     if (config.objcInterfaces.rename(_decl(node)) case final rename?) {
       node.name = rename;
     }
+    if (config.interfaceModule(_decl(node)) case final module?) {
+      node.module = module;
+    }
   }
 
   @override
@@ -1402,6 +1397,9 @@ final class YamlConfigAstVisitor extends public_ast.Visitor {
     node.isIncluded = config.objcProtocols.shouldInclude(_decl(node));
     if (config.objcProtocols.rename(_decl(node)) case final rename?) {
       node.name = rename;
+    }
+    if (config.protocolModule(_decl(node)) case final module?) {
+      node.module = module;
     }
   }